+++ R/Object.R
@@ -15,10 +15,39 @@
 # limitations under the License.
 
 #' Andromeda class.
+#' 
+#' @description 
+#' The Andromeda class is an S4 object.
+#' 
+#' This class provides the ability to work with data objects in R that are too large to fit in memory. Instead, 
+#' these objects are stored on disk. This is slower than working from memory, but may be the only viable option. 
+#' 
+#' @section Tables:
+#' An `Andromeda` object has zero, one or more tables. The list of table names can be retrieved using the \code{names} 
+#' method. Tables can be accessed using the dollar sign syntax, e.g. \code{andromeda$myTable}.
 #'
-#' @keywords internal
-#' @export
+#' 
+#' @section Permanence:
+#' 
+#' To mimic the behavior of in-memory objects, when working with data in Andromeda the data is stored in a 
+#' temporary location on the disk. You can modify the data as you can see fit, and when needed can save the data 
+#' to a permanent location. Later this data can be loaded to a temporary location again and be read and modified, 
+#' while keeping the saved data as is.
+#' 
+#' @section Inheritance:
+#' 
+#' The Andromeda inherits directly from \code{SQLiteConnection}. As such, it can be used as if it is a \code{SQLiteConnection}. 
+#' \code{RSQLite} is an R wrapper around SQLite, a low-weight but very powerful single-user SQL database that can run 
+#' from a single file on the local file system.
+#' 
+#' @name Andromeda-class
+#' @seealso \code{\link{andromeda}}
+NULL
+
+#' Andromeda class.
+#'
 #' @import RSQLite
+#' @export
 setClass("Andromeda", contains = "SQLiteConnection")
 
 #' Create an Andromeda object
@@ -33,6 +62,9 @@ setClass("Andromeda", contains = "SQLiteConnection")
 #'
 #' @details
 #' Valid objects are data frames, Andromeda tables, or any other dply table.
+#' 
+#' @return 
+#' Returns an \code{\link{Andromeda-class}} object.
 #'
 #' @examples
 #' andr <- andromeda(cars = cars, iris = iris)