+++ NEWS.md
@@ -5,6 +5,8 @@ Changes
 
 - Minor edits to documentation in preparation for submission to CRAN.
 
+- Changing compression level when saving: 10-fold reduction in compression time at the cost of 10% larger file size.
+
 Andromeda 0.1.0
 ================