+++ NEWS.md
@@ -7,6 +7,10 @@ Changes
 
 - Changing compression level when saving: 10-fold reduction in compression time at the cost of 10% larger file size.
 
+- Increasing batch size from 10,000 to 100,000 to increase speed.
+
+- Turning of SQLIte journal to increase speed.
+
 Andromeda 0.1.0
 ================